Refactor Section into component, logout form
